backend/requirements.txt
streamlit
assemblyai
python-dotenv